refactor: extract PDF canonical node composition
